fix: widen Label::ref_count past uint8_t

+++ src/rcspp/label/label.hpp
@@ -4,6 +4,7 @@
 #pragma once
 
 #include <concepts>
+#include <cstdint>
 #include <iostream>
 #include <memory>
 #include <utility>
@@ -83,8 +84,11 @@ class Label {
 
         // Predecessor label set at extension time; valid as long as ref_count keeps it pinned.
         Label<ResourceType>* prev_label = nullptr;
-        // Number of alive successors that reference this label as their predecessor.
-        uint8_t ref_count = 0;
+        // Number of alive successors that reference this label as their predecessor. 32 bits so a
+        // high-out-degree node (e.g. a dense VRP pricing graph that extends one label to many
+        // hundreds or thousands of successors) cannot overflow the count, which would make
+        // release_with_ref_count() free a still-referenced predecessor.
+        uint32_t ref_count = 0;
         // True when the algorithm wanted to release this label but ref_count was > 0.
         bool pending_release = false;
